Move health autoconfig into health package
This commit moves the health autoconfiguration into its own
package as well as moves the ActuatorHealthAdapter into its
own class.
