+++ spring-grpc-spring-boot-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,7 +1,7 @@
 org.springframework.grpc.autoconfigure.client.GrpcClientAutoConfiguration
 org.springframework.grpc.autoconfigure.server.GrpcServerFactoryAutoConfiguration
 org.springframework.grpc.autoconfigure.server.GrpcServerAutoConfiguration
-org.springframework.grpc.autoconfigure.server.GrpcServerHealthAutoConfiguration
+org.springframework.grpc.autoconfigure.server.health.GrpcServerHealthAutoConfiguration
 org.springframework.grpc.autoconfigure.server.GrpcServerObservationAutoConfiguration
 org.springframework.grpc.autoconfigure.server.GrpcServerReflectionAutoConfiguration
 org.springframework.grpc.autoconfigure.server.exception.GrpcExceptionHandlerAutoConfiguration
